refactor(planner): Clean up planner files & make small UI changes
